requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
plotly
xlsxwriter
scikit-learn
